requirements.txt
crewai
crewai[tools]
python-dotenv
openai
langchain_openai
langchain_groq
pydub